rename: 유효성 검사 함수들 파일 이름 변경
