+++ src/app/pockets/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import usePocketStore from "@/store/pocket";
 import { fetchLuckyPouches } from "@/api/api-form";
 import Notice from "@/components/modal/notice";
-import { validatePouches, Pouch } from "@/utils/validation/validationPouch";
+import { validatePouches, Pouch } from "@/utils/validation/validate-pouches";
 import { getPouch } from "@/utils/images/domain";
 import Header from "@/components/header/header";
 import Logo from "@/components/header/logo";
