+++ src/app/result/page.tsx
@@ -6,7 +6,7 @@ import React, { useEffect, useState } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { fetchResultData, ResultData } from "../../api/api-result";
 import ResultModal from "@/components/modal/result-modal";
-//import {ValidateResult} from "@/components/result/resultValidation";
+// import { validateResult } from "@/utils/validation/validate-result";
 import { fetchRightAnswers, fetchWrongAnswers } from "@/api/api-result-data";
 import useResultStore from "@/store/result";
 import Header from "@/components/header/header";
