+++ src/utils/validation/validate-result.ts
